// src/enums/Image.enum.ts
export enum GetImagesEnum {
  FAILED_TO_GET_IMAGES = "FAILED_TO_GET_IMAGES",
  USER_IMAGES = "USER_IMAGES",
}

export enum GetImageByIdEnum {
  FAILED_TO_GET_IMAGE = "FAILED_TO_GET_IMAGE",
  USER_IMAGE = "USER_IMAGE",
}

export enum CreateImagesEnum {
  FAILED_TO_CREATE_IMAGES = "FAILED_TO_CREATE_IMAGES",
  IMAGE_CREATED = "IMAGE_CREATED",
  DAILY_COUNT_EXCEDDED = "DAILY_COUNT_EXCEDDED",
}

export enum GetTotalViewsCountEnum {
  FAILED_TO_GET_VIEWS = "FAILED_TO_GET_VIEWS",
  FOUND_VIEWS = "FOUND_USER_DETAIL",
}

export enum GetImagesByTagsEnum {
  IMAGE_FOUND = "IMAGE_FOUND",
  FAILED_TO_GET_IMAGES = "FAILED_TO_GET_IMAGES",
}

export enum GetAllImagesEnum {
  IMAGE_FOUND = "IMAGE_FOUND",
  FAILED_TO_GET_ALL_IMAGES = "FAILED_TO_GET_ALL_IMAGES",
}